add authorization on edit and delete page

+++ project2/controllers/feed.js
@@ -125,6 +125,11 @@ exports.updatePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorized!");
+        error.statusCode = 403;
+        throw error;
+      }
       if (imageUrl !== post.imageUrl) {
         clearImage(post.imageUrl);
       }
@@ -153,6 +158,11 @@ exports.deletePost = (req, res, next) => {
         error.statusCode = 404;
         throw error;
       }
+      if (post.creator.toString() !== req.userId) {
+        const error = new Error("Not authorization.");
+        error.statusCode = 403;
+        throw error;
+      }
       // check if logged in
       clearImage(post.imageUrl);
       return Post.findByIdAndRemove(postId);
@@ -165,7 +175,7 @@ exports.deletePost = (req, res, next) => {
       if (!err.statusCode) {
         err.statusCode = 500;
       }
-      nex(err);
+      next(err);
     });
 };
